fixes for compiler warnings on your new batch argument (similar as last push that I made) (#194)

+++ src/starspace.cpp
@@ -336,7 +336,7 @@ Metrics StarSpace::evaluateOne(
   for (unsigned int i = 0; i < baseDocVectors_.size(); i++) {
     // in the case basedoc labels are not provided, all labels become basedoc,
     // and we skip the correct label for comparison.
-    if ((args_->basedoc.empty()) && (i == rhs[0].first - dict_->nwords())) {
+    if ((args_->basedoc.empty()) && ((int)i == rhs[0].first - dict_->nwords())) {
       continue;
     }
     auto cur_score = model_->similarity(lhsM, baseDocVectors_[i]);
